Improve stringify() code and tests
stringify() now uses more descriptive variables and replaces the =,+=,+
operators with just = to make the code a little easier to take in.

The tests now include a string array and a long entry to see if it
prints out well. Mixing types does not work unless I edit comparison,
but I shall leave for simplicity and will just say that the
functionality of the Tree class requires the same type.