Create insert() of Tree class
Uses Node comparison for easier readability.

Loop checks for duplicate to check if node already inserted, either from
before or just now. We remove duplicates as per the instructions.

Symbols and send() are used to avoid the use of if/else conditionals and
instead only include one ternary operator. In my opinion the readability
is not impacted too much by this and lines of code are reduced a lot.

The Tree object is returned in every case to allow a chain of messages
to be possible, like chaining multiple insert messages. As an
alternative insert() could accept an array argument but this goes
against the instructions for this method specifically.