Create find() method in Tree class
Did this by creating a reusable private method locate() which returns
the node if found or the closest node if not found (and nil for an empty
tree). This can be reused for insert, delete, and maybe more!

I also added a to_s method to the Node class to make printing nodes
easier.