Create Parent; find(), insert(); get, set methods
I did a lot in this commit just to get my thoughts down and experiment!

I realized that I would need to find the parent node for every search
type method (the other half are traverse type methods). Rather than
repeating the same code I decided to write it in a method that can be
reused. This method returned the parent node but the caller still often
needed to figure out the direction again and this is just more repeated
code. So the direction also needs to be returned, but in what format?
Using symbols one can use the send method to call .left, .left=, etc.
This could be returned in an array, but I was realizing that what the
method would return is effectively an object with data and methods! So I
decided to create the Parent class which finds the parent of the search
node with the ability to read/write its child. I then decided to absorb
the original method I was reusing into Parent since it was using only
Parent methods after a certain point and felt like its responsibility.

I then rewrote the find and insert methods to see what they looked like
and how the Parent object would work within them. I hope that the Parent
class creates both DRY and readable code.

I have still to test the new find and insert methods but decided to make
this commit about Parent.

I also made the small addition of using getter and setter methods for
every instance variable (making the setters mostly private). This is
better than using the instance variable directly; it makes the code more
easily changeable (one change in how a variable is read will create
change everywhere it is used!). I started in Parent and thought I might
as well continue everywhere so I included that in this commit too.